♻️ refactor : 언마운트시 이미지 삭제  #70
